// src/shared/api/generated.ts
export type components = {
	schemas: {
// ...
		PipelineRunResponse: {
			/** The body of the resource. */
			body?: components["schemas"]["PipelineRunResponseBody"] | null;
			/** The metadata related to this resource. */
			metadata?: components["schemas"]["PipelineRunResponseMetadata"] | null;
			/** The resources related to this resource. */
			resources?: components["schemas"]["PipelineRunResponseResources"] | null;
			/**
			 * The unique resource id.
			 * Format: uuid
			 */
			id: string;
			/**
			 * Permission Denied
			 * @default false
			 */
			permission_denied: boolean;
			/** The name of the pipeline run. */
			name: string;
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
};